Add test and fix a bunch of code review issues
